Shimcache and amcache parsing to their own files
